Fix unit test intgral types

+++ src/torrent/choker.rs
@@ -196,7 +196,7 @@ mod tests {
         let mut h = HashMap::new();
         assert_eq!(c.update_upload(&mut h).is_none(), true);
         for i in 0..6 {
-            let mut p = Peer::test_from_stats(i, i as u64, 6 - i as u64);
+            let mut p = Peer::test_from_stats(i, i as u32, 6 - i as u32);
             c.add_peer(&mut p);
             h.insert(i, p);
         }
@@ -213,7 +213,7 @@ mod tests {
         let mut h = HashMap::new();
         assert_eq!(c.update_download(&mut h).is_none(), true);
         for i in 0..6 {
-            let mut p = Peer::test_from_stats(i, 6 - i as u64, i as u64);
+            let mut p = Peer::test_from_stats(i, 6 - i as u32, i as u32);
             c.add_peer(&mut p);
             h.insert(i, p);
         }
