+++ src/torrent/peer/mod.rs
@@ -142,8 +142,8 @@ impl Status {
 impl Peer<cio::test::TCIO> {
     pub fn test(
         id: usize,
-        uploaded: u64,
-        downloaded: u64,
+        uploaded: u32,
+        downloaded: u32,
         queued: u16,
         pieces: Bitfield,
     ) -> Peer<cio::test::TCIO> {
@@ -171,7 +171,7 @@ impl Peer<cio::test::TCIO> {
         Peer::test(id, 0, 0, 0, pieces)
     }
 
-    pub fn test_from_stats(id: usize, ul: u64, dl: u64) -> Peer<cio::test::TCIO> {
+    pub fn test_from_stats(id: usize, ul: u32, dl: u32) -> Peer<cio::test::TCIO> {
         Peer::test(id, ul, dl, 0, Bitfield::new(4))
     }
